third_party/WebKit/LayoutTests/virtual/off-main-thread-fetch/http/tests/misc/resource-timing-sizes-cors-preflight-worker-expected.txt

This is a testharness.js-based test.
FAIL PerformanceResourceTiming sizes Fetch with preflight test assert_equals: Wrong number of entries expected 3 but got 0
Harness: the test ran to completion.
